requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.14
all-clip==1.2.0
aniso8601==9.0.1
annotated-types==0.7.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==24.2.0
autofaiss==2.17.0
blinker==1.8.2
braceexpand==0.1.7
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
clip-anytorch==2.6.0
clip-retrieval==2.44.0
cloudpickle==3.0.0
codetiming==1.4.0
dataclasses==0.6
docker-pycreds==0.4.0
docutils==0.21.2
embedding-reader==1.7.0
eval_type_backport==0.2.0
ExifRead-nocycle==3.0.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
fire==0.5.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
h5py==3.11.0
huggingface-hub==0.24.6
idna==3.7
imageio==2.35.1
img2dataset==1.45.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
line_profiler==4.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multilingual-clip==1.0.10
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.26.1
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
prometheus_client==0.20.0
protobuf==4.25.4
psutil==6.0.0
pyarrow==14.0.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.8.0
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.11.4
sentence-transformers==2.7.0
sentry-sdk==2.13.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
statistics==1.0.3.5
sympy==1.13.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.8.10
timm==1.0.8
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.1
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
wandb==0.16.6
wcwidth==0.2.13
webdataset==0.2.96
Werkzeug==3.0.4
yarl==1.9.4
